fix(http): enable TCP keepalive on outbound transports
The HTTP transports created at three hot-path call sites
(http_json_rpc_client, proxy_pool_registry, dynamodb shared clients)
did not set a custom DialContext, so they used the zero-value
net.Dialer. With KeepAlive=0, Go skips SO_KEEPALIVE / TCP_KEEPIDLE
setsockopts and the kernel falls back to its system default
(tcp_keepalive_time=7200s on Linux). That means a wedged outbound
TCP flow — where the kernel still believes the connection is alive
but no data is flowing — is invisible to kernel-level cleanup for
two hours.

When such a wedge happens, the standard application-level
defenses (http.Client.Timeout, Transport.ResponseHeaderTimeout,
context.WithDeadline) all rely on conn.Close() interrupting the
in-flight recv() syscall. If the kernel TCP state is in a
half-closed-but-not-RSTed mode, that close doesn't wake the
syscall promptly and the call hangs.

Setting KeepAlive=15s makes the kernel send TCP keepalive probes
every 15s. Three missed probes (~45s) trigger an RST; the
recv() syscall returns; the application-level timeouts then
fire cleanly. This is exactly what Go's own http.DefaultTransport
does (KeepAlive=30s) and is the standard pattern for production
outbound HTTP clients.

Also refactor: move the BDS gRPC-specific callBounded primitive
to util/bounded_call.go where it can be referenced generically;
the clients package keeps thin aliases so existing call sites
(and the BDS resilience tests) continue to work unchanged.

Verification: make test-fast green across all 12 packages, go vet
clean, go build clean.
