+++ clients/http_json_rpc_client.go
@@ -6,6 +6,7 @@ import (
 	"errors"
 	"fmt"
 	"io"
+	"net"
 	"net/http"
 	"net/url"
 	"strings"
@@ -102,8 +103,19 @@ func NewGenericHttpJsonRpcClient(
 	}
 
 	// Default fallback transport (no proxy)
-	// Optimized for high-latency, high-RPS scenarios to prevent connection churn
+	// Optimized for high-latency, high-RPS scenarios to prevent connection churn.
+	//
+	// DialContext sets KeepAlive so the kernel sends TCP keepalive probes
+	// every 15s. Without this, Go falls back to the zero-value Dialer which
+	// uses the OS default tcp_keepalive_time (7200s = 2h on Linux), so
+	// wedged outbound flows can't be detected at the kernel layer. With
+	// 15s probes, a flow that loses connectivity is killed in ~45s
+	// (3 missed probes), which lets the application-level timeouts fire.
 	transport := &http.Transport{
+		DialContext: (&net.Dialer{
+			Timeout:   10 * time.Second,
+			KeepAlive: 15 * time.Second,
+		}).DialContext,
 		MaxIdleConns:          1024,
 		MaxIdleConnsPerHost:   256,
 		MaxConnsPerHost:       0, // Unlimited active connections (prevents bottleneck)
